[수정] 불필요한 파일 삭제 및 리팩토링
